Ok so we've worked out why the PNGTest example doesn't work on QEMU, it is to do with the color format of the QEMU framebuffer being different to the Raspberry Pi.
You can make it work easily by finding this line in your PNGTest.lpr

Code: Select all

      FBFormat := FrameProps.Format;

and changing it to this instead:

Code: Select all

      FBFormat := COLOR_FORMAT_DEFAULT;

That works because the QEMU framebuffer is COLOR_FORMAT_URGB32 whereas the Raspberry Pi is COLOR_FORMAT_ARGB32 so the only difference is the alpha channel which isn't important in this case.

Ideally the uCanvas unit should be able to work with different color formats for different platforms and as we said above we'll see if we can have a look at what changes are needed to do that.

In the meantime we've created a fork of the PNG project on our GitHub and included these changes, you can find it here.
Ultibo.org | Make something amazing
https://ultibo.org